+++ lib/pages/AXPage.dart
@@ -1,4 +1,88 @@
+import 'dart:math';
+import 'dart:html' as html;
+
+
+// flutter format .
+
+import 'package:ae_dapp/service/colors.dart';
+import 'package:url_launcher/url_launcher.dart';
 import 'package:flutter/material.dart';
+import 'package:toggle_switch/toggle_switch.dart';
+import 'package:ae_dapp/style/Style.dart';
+
+_launchURL() async {
+  if (await canLaunch(url)) {
+    await launch(url);
+  } else {
+    throw 'Could not launch $url';
+  }
+}
+
+
+Widget _buildPopupDialog(BuildContext context) {
+  return new AlertDialog(
+    
+    title: Text('Connect to a wallet',
+    textAlign: TextAlign.center,
+    style: TextStyle(color: Colors.grey[400],
+    fontFamily: 'OpenSans',
+    )
+    ),
+    content: new Column(
+      mainAxisSize: MainAxisSize.min,
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: <Widget>[
+        Padding(
+          padding: EdgeInsets.fromLTRB(0, 0, 0, 5),
+          child: ElevatedButton(
+            child: Align(
+              alignment: Alignment.centerLeft,
+                child: Padding(
+                  padding: EdgeInsets.fromLTRB(5, 0, 0, 0),
+                  child: Text('Install Metamask'),
+                  ),
+              ),
+            onPressed: () {},
+            style: walletButton,
+          ),
+        ),
+
+        Padding(
+          padding: EdgeInsets.fromLTRB(0, 0, 0, 5),
+          child: ElevatedButton(
+            child: Align(
+              alignment: Alignment.centerLeft,
+                child: Padding(
+                  padding: EdgeInsets.fromLTRB(5, 0, 0, 0),
+                  child: Text('WalletConnect'),
+                  ),
+              ),
+            onPressed: () {},
+            style: walletButton,
+          ),
+        ),
+
+        Padding(
+          padding: EdgeInsets.fromLTRB(0, 0, 0, 5),
+          child: ElevatedButton(
+            child: Align(
+              alignment: Alignment.centerLeft,
+                child: Padding(
+                  padding: EdgeInsets.fromLTRB(5, 0, 0, 0),
+                  child: Text('Coinbase Wallet'),
+                  ),
+              ),
+            onPressed: () {},
+            style: walletButton,
+          ),
+        ),
+
+
+      ],
+    ),
+    
+  );
+}
 
 class AXPage extends StatefulWidget {
   @override
@@ -6,258 +90,422 @@ class AXPage extends StatefulWidget {
 }
 
 class _AXState extends State<AXPage> {
-  Widget build(BuildContext context) {
-    double lgTxSize = 52;
-    double smTxSize = 20;
-    // Column
-    // Row
-    // AX
-    // X
-    // Container
-    // Row
-    // Column
-    // My Account
-    // ...
-    // Column
-    // My Awards
-    // ...
-    // Container
-    // Row
-    // Column
-    // Buy AX
-    // Stake AX
-    // Column
-    // Claim Rewards
-    // Unstake AX
+  TextEditingController stakeController = new TextEditingController();
 
+  // Actionable
+  Future<void> buyAX() async {}
+  Future<void> stakeAX() async {}
+  Future<void> claimRewards() async {}
+  Future<Widget> unstakeAX() async {
+    return Text("Unclaimed Rewards: ");
+  }
+
+  var currentTotal = 0;
+  var numNewAX = 0;
+
+  // Viewable
+  Future<int> totalBalance() async {
+    currentTotal = new Random().nextInt(800);
+    return currentTotal;
+  }
+
+  Future<int> totalBalanceStake(int numNewAX) async {
+    return currentTotal + numNewAX;
+  }
+
+  Future<String> availableBalance() async {
+    return "0";
+  }
+
+  Future<int> stakedAX() async {
+    return Random().nextInt(200);
+  }
+
+  Future<int> getAPY() async {
+    return Random().nextInt(200);
+  }
+
+  Future<int> rewardsEarned() async {
+    return new Random().nextInt(200);
+  }
+
+  Future<int> UnclaimedRewards() async {
+    return new Random(300).nextInt(200);
+  }
+
+  Widget build(BuildContext context) {
+    
     return Scaffold(
         body: Stack(
       children: <Widget>[
+        // background image
         Image(
           image: AssetImage("assets/images/background.jpeg"),
           width: MediaQuery.of(context).size.width,
           height: MediaQuery.of(context).size.height,
           fit: BoxFit.cover,
         ),
-        Column(children: <Widget>[
-          Align(
-              alignment: Alignment.topRight,
-              child: Padding(
-                padding: const EdgeInsets.only(right: 20),
-                child: Image(
-                  width: 80,
-                  height: 80,
-                  image: AssetImage("assets/images/x.png"),
-                ),
-              )),
-          Text("AX",
-              style: TextStyle(
-                fontFamily: 'OpenSans',
-                fontSize: lgTxSize,
-                fontWeight: FontWeight.w600,
-                fontStyle: FontStyle.italic,
-              )),
-          Padding(
-              padding: EdgeInsets.fromLTRB(30, 30, 30, 0),
-              child: Container(
-                decoration: BoxDecoration(
+
+        // upper right logo
+        Align(
+            alignment: Alignment.topRight,
+            child: Padding(
+              padding: const EdgeInsets.only(right: 20),
+              child: Image(
+                width: 80,
+                height: 80,
+                image: AssetImage("assets/images/x.png"),
+              ),
+            )),
+
+        // header
+        Container(
+            margin: EdgeInsets.fromLTRB(0, 100, 0, 0),
+            child: Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                crossAxisAlignment: CrossAxisAlignment.center,
+                children: [
+                  Text("AX",
+                      style: TextStyle(
+                        fontFamily: 'OpenSans',
+                        fontSize: lgTxSize,
+                        fontWeight: FontWeight.w600,
+                        fontStyle: FontStyle.italic,
+                      )),
+                ])),
+
+        // main user area
+        Container(
+          child: Row(
+          mainAxisAlignment: MainAxisAlignment.center,
+          crossAxisAlignment: CrossAxisAlignment.center,
+          children: [
+
+            // my account info
+            Column(
+              children: [
+                // Main Left Area
+                Container(
+                    margin: EdgeInsets.fromLTRB(0, 175, 0, 0),
+                    alignment: Alignment.center,
+                    height: 500,
+                    width: 400,
                     color: Colors.grey[900],
-                    borderRadius: BorderRadius.all(Radius.circular(20))),
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                  children: <Widget>[
-                    Column(
-                      children: <Widget>[
-                        Padding(
-                            padding: EdgeInsets.symmetric(vertical: 10),
-                            child: Text("My Account",
-                                style: TextStyle(
+                    child: Row(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      crossAxisAlignment: CrossAxisAlignment.center,
+
+                      // Main Left Area
+                      children: [
+                        Column(
+                          children: [
+                            // My Account
+                            Padding(
+                              padding: EdgeInsets.fromLTRB(0, 45, 0, 20),
+                              child: Text(
+                                'My Account',
+                                  style: TextStyle(
+                                    fontSize: 25,
                                     fontFamily: 'OpenSans',
-                                    fontSize: lgTxSize,
-                                    fontWeight: FontWeight.w600))),
-                        Text("AX Total Balance: ",
-                            style: TextStyle(
-                              fontFamily: 'OpenSans',
-                              fontSize: smTxSize,
-                            )),
-                        Padding(
-                            padding: EdgeInsets.symmetric(vertical: 5),
-                            child: Text("AX Available Balance: ",
-                                style: TextStyle(
-                                  fontFamily: 'OpenSans',
-                                  fontSize: smTxSize,
-                                ))),
-                        Padding(
-                            padding: EdgeInsets.only(bottom: 10),
-                            child: Text("AX Staked: ",
-                                style: TextStyle(
-                                  fontFamily: 'OpenSans',
-                                  fontSize: smTxSize,
-                                ))),
-                      ],
-                    ),
-                    Column(
-                      children: <Widget>[
-                        Padding(
-                            padding: EdgeInsets.symmetric(vertical: 10),
-                            child: Text("My Rewards",
-                                style: TextStyle(
-                                  fontFamily: 'OpenSans',
-                                  fontSize: lgTxSize,
-                                  fontWeight: FontWeight.w600,
-                                ))),
-                        Text("APY: ",
-                            style: TextStyle(
-                              fontFamily: 'OpenSans',
-                              fontSize: smTxSize,
-                            )),
-                        Padding(
-                            padding: EdgeInsets.symmetric(vertical: 5),
-                            child: Text("Rewards Earned: ",
-                                style: TextStyle(
-                                  fontFamily: 'OpenSans',
-                                  fontSize: smTxSize,
-                                ))),
-                        Padding(
-                            padding: EdgeInsets.only(bottom: 10),
-                            child: Text("Unclaimed Rewards: ",
-                                style: TextStyle(
-                                  fontFamily: 'OpenSans',
-                                  fontSize: smTxSize,
-                                ))),
-                      ],
-                    )
-                  ],
-                ),
-                // color: Colors.grey[900],
-              )),
-          Padding(
-              padding: EdgeInsets.symmetric(horizontal: 30),
-              child: Container(
-                  decoration: BoxDecoration(
-                      color: Colors.grey[800],
-                      borderRadius: BorderRadius.all(Radius.circular(20))),
-                  child: Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                      children: <Widget>[
-                        Column(
-                          children: [
+                                    fontWeight: FontWeight.w600,
+                                    fontStyle: FontStyle.italic
+                                  )
+                                ),
+                            ),
+
+                            // Staked AX
                             Padding(
-                                padding: EdgeInsets.symmetric(vertical: 20),
-                                child: ConstrainedBox(
-                                    constraints: BoxConstraints.tightFor(
-                                        width: 250, height: 55),
-                                    child: ElevatedButton(
-                                      child: Text("BUY AX",
-                                          style: TextStyle(
-                                            fontFamily: 'OpenSans',
-                                            fontSize: smTxSize,
-                                            fontWeight: FontWeight.w600,
-                                            color: Colors.white,
-                                          )),
-                                      onPressed: () {},
-                                      style: ButtonStyle(
-                                          backgroundColor:
-                                              MaterialStateProperty.all<Color>(
-                                                  Colors.amber[400]!),
-                                          shape: MaterialStateProperty.all<
-                                                  RoundedRectangleBorder>(
-                                              RoundedRectangleBorder(
-                                            borderRadius:
-                                                BorderRadius.circular(18.0),
-                                            side: BorderSide(
-                                                color: Colors.amber[400]!),
-                                          ))),
-                                    ))),
+                              padding: EdgeInsets.fromLTRB(0, 0, 0, 5),
+                              child: Text(
+                                'Staked AX: ',
+                                style: TextStyle(
+                                  fontSize: smTxSize,
+                                  fontFamily: 'OpenSans',
+                                  fontWeight: FontWeight.w400,
+                                ),
+                              ),
+                            ),
+
+                            // Available AX
                             Padding(
-                                padding: EdgeInsets.only(bottom: 20),
-                                child: ConstrainedBox(
-                                    constraints: BoxConstraints.tightFor(
-                                        width: 250, height: 55),
-                                    child: ElevatedButton(
-                                      child: Text("STAKE AX",
-                                          style: TextStyle(
-                                            fontFamily: 'OpenSans',
-                                            fontSize: smTxSize,
-                                            fontWeight: FontWeight.w600,
-                                            color: Colors.amber[400],
-                                          )),
-                                      onPressed: () {},
-                                      style: ButtonStyle(
-                                          backgroundColor:
-                                              MaterialStateProperty.all<Color>(
-                                                  Colors.grey[800]!),
-                                          shape: MaterialStateProperty.all<
-                                                  RoundedRectangleBorder>(
-                                              RoundedRectangleBorder(
-                                            borderRadius:
-                                                BorderRadius.circular(18.0),
-                                            side: BorderSide(
-                                                color: Colors.amber[400]!),
-                                          ))),
-                                    ))),
-                          ],
-                        ),
-                        Column(
-                          children: [
+                              padding: EdgeInsets.fromLTRB(0, 0, 0, 5),
+                              child: Text(
+                                'Available AX: ',
+                                style: TextStyle(
+                                  fontSize: smTxSize,
+                                  fontFamily: 'OpenSans',
+                                  fontWeight: FontWeight.w400,
+                                ),
+                              ),
+                            ),
+
+                            // Total AX
                             Padding(
-                                padding: EdgeInsets.symmetric(vertical: 20),
-                                child: ConstrainedBox(
-                                    constraints: BoxConstraints.tightFor(
-                                        width: 250, height: 55),
-                                    child: ElevatedButton(
-                                      child: Text("CLAIM REWARDS",
-                                          style: TextStyle(
-                                            fontFamily: 'OpenSans',
-                                            fontSize: smTxSize,
-                                            fontWeight: FontWeight.w600,
-                                            color: Colors.white,
-                                          )),
-                                      onPressed: () {},
-                                      style: ButtonStyle(
-                                          backgroundColor:
-                                              MaterialStateProperty.all<Color>(
-                                                  Colors.amber[400]!),
-                                          shape: MaterialStateProperty.all<
-                                                  RoundedRectangleBorder>(
-                                              RoundedRectangleBorder(
-                                            borderRadius:
-                                                BorderRadius.circular(18.0),
-                                            side: BorderSide(
-                                                color: Colors.amber[400]!),
-                                          ))),
-                                    ))),
+                              padding: EdgeInsets.fromLTRB(0, 0, 0, 5),
+                              child: Text(
+                                'Total AX: ',
+                                style: TextStyle(
+                                  fontSize: smTxSize,
+                                  fontFamily: 'OpenSans',
+                                  fontWeight: FontWeight.w400,
+                                ),
+                              ),
+                            ),
+
+                            // My Rewards
                             Padding(
-                                padding: EdgeInsets.only(bottom: 20),
-                                child: ConstrainedBox(
-                                    constraints: BoxConstraints.tightFor(
-                                        width: 250, height: 55),
-                                    child: ElevatedButton(
-                                      child: Text("UNSTAKE AX",
-                                          style: TextStyle(
-                                            fontFamily: 'OpenSans',
-                                            fontSize: smTxSize,
-                                            fontWeight: FontWeight.w600,
-                                            color: Colors.amber[400],
-                                          )),
-                                      onPressed: () {},
-                                      style: ButtonStyle(
-                                          backgroundColor:
-                                              MaterialStateProperty.all<Color>(
-                                                  Colors.grey[800]!),
-                                          shape: MaterialStateProperty.all<
-                                                  RoundedRectangleBorder>(
-                                              RoundedRectangleBorder(
-                                            borderRadius:
-                                                BorderRadius.circular(18.0),
-                                            side: BorderSide(
-                                                color: Colors.amber[400]!),
-                                          ))),
-                                    ))),
+                              padding: EdgeInsets.fromLTRB(0, 20, 0, 20),
+                              child: Text('My Rewards',
+                                  style: TextStyle(
+                                    fontSize: 25,
+                                    fontFamily: 'OpenSans',
+                                    fontWeight: FontWeight.w600,
+                                    fontStyle: FontStyle.italic
+                                  )),
+                            ),
+
+                            // APY
+                            Padding(
+                              padding: EdgeInsets.fromLTRB(0, 0, 0, 5),
+                              child: Text(
+                                'APY: ',
+                                style: TextStyle(
+                                  fontSize: smTxSize,
+                                  fontFamily: 'OpenSans',
+                                  fontWeight: FontWeight.w400,
+                                ),
+                              ),
+                            ),
+
+                            // Rewards Earned
+                            Padding(
+                              padding: EdgeInsets.fromLTRB(0, 0, 0, 5),
+                              child: Text(
+                                'Rewards Earned: ',
+                                style: TextStyle(
+                                  fontSize: smTxSize,
+                                  fontFamily: 'OpenSans',
+                                  fontWeight: FontWeight.w400,
+                                ),
+                              ),
+                            ),
+
+                            // Unclaimed Rewards
+                            Padding(
+                              padding: EdgeInsets.fromLTRB(0, 0, 0, 0),
+                              child: Text(
+                                'Unclaimed Rewards: ',
+                                style: TextStyle(
+                                  fontSize: smTxSize,
+                                  fontFamily: 'OpenSans',
+                                  fontWeight: FontWeight.w400,
+                                ),
+                              ),
+                            ),
+                            //Claim rewards button
+                            Padding(
+                              padding: EdgeInsets.fromLTRB(0, 20, 0, 0),
+                              child: SizedBox(
+                                child: ElevatedButton(
+                                  child: Text('CLAIM REWARDS'),
+                                  onPressed: () {},
+                                  style: claimButton,
+                                ),
+                                height: 50,
+                                width: 250,
+                              )),
+                            //Buy AX button
+                            Padding(
+                              padding: EdgeInsets.fromLTRB(0, 10, 0, 0),
+                              child: SizedBox(
+                                child: ElevatedButton(
+                                  child: Text('BUY AX'),
+                                  onPressed: _launchURL,
+                                  style: claimButton,
+                                ),
+                                height: 50,
+                                width: 250,
+                              )),
                           ],
                         )
-                      ]))),
-        ])
+                      ],
+                    )),
+              ],
+            ),
+            
+            //Main right area
+            Column(
+              children: [
+                //Main Right Area
+                Container(
+                    margin: EdgeInsets.fromLTRB(0, 175, 0, 0),
+                    alignment: Alignment.center,
+                    height: 500,
+                    width: 500,
+                    color: Colors.grey[800],
+
+                    // My Liquidity Area
+                    child: Row(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      crossAxisAlignment: CrossAxisAlignment.center,
+                      children: [
+                        Container(
+                          
+                          child: Column(
+                          children: [
+                            // My Liquidity
+                            Padding(
+                                padding: EdgeInsets.fromLTRB(0, 45, 0, 10),
+                                child: Text('My Liquidity',
+                                    style: TextStyle(
+                                      fontSize: mdTxSize,
+                                      fontFamily: 'OpenSans',
+                                      fontWeight: FontWeight.w600,
+                                      fontStyle: FontStyle.italic
+                                    ))),
+
+                            // adding note
+                            Container(
+                                child: Column(
+                              children: [
+                                // Add liquidity tag
+                                Padding(
+                                    padding: EdgeInsets.fromLTRB(0, 0, 0, 0),
+                                    child: Text('Add liquidity to stake AX',
+                                        style: TextStyle(
+                                          fontSize: xsTxSize,
+                                          fontFamily: 'OpenSans',
+                                          fontWeight: FontWeight.w200,
+                                        ))),
+
+                                // remove note
+                                Padding(
+                                    padding: EdgeInsets.fromLTRB(0, 0, 0, 40),
+                                    child:
+                                        Text('Remove liquidity to receive AX',
+                                            style: TextStyle(
+                                              fontSize: xsTxSize,
+                                              fontFamily: 'OpenSans',
+                                              fontWeight: FontWeight.w200,
+                                            ))),
+
+                                Container(
+                                  child: Row(
+                                  children: [
+                                    Column(children: [
+                                      ToggleSwitch(
+                                        minWidth: 75.0,
+                                        minHeight: 30.0,
+                                        fontSize: xsTxSize,
+                                        activeBgColor: [(Colors.grey[800])!],
+                                        activeFgColor: Colors.amber[600],
+                                        inactiveBgColor: Colors.grey[800],
+                                        inactiveFgColor: Colors.white,
+                                        initialLabelIndex: 0,
+                                        totalSwitches: 2,
+                                        labels: ['STAKE', 'UNSTAKE'],
+                                        onToggle: (index) {
+                                        },
+                                      ),
+                                    ],),
+
+                                  Column(children: [
+                                      ToggleSwitch(
+                                        borderColor: [(Colors.grey[800])!],
+                                        minWidth: 50.0,
+                                        minHeight: 30.0,
+                                        fontSize: xsTxSize,
+                                        activeBgColor: [(Colors.grey[800])!],
+                                        activeFgColor: Colors.amber[600],
+                                        inactiveBgColor: Colors.grey[800],
+                                        inactiveFgColor: Colors.white,
+                                        initialLabelIndex: 0,
+                                        totalSwitches: 4,
+                                        labels: ['25%', '50%', '75%', 'MAX'],
+                                        onToggle: (index) {
+                                        },
+                                      ),
+                                    ],),
+
+                                    
+                                  ],
+                                )),
+
+                                // text field box
+                                Padding(
+                                  padding: EdgeInsets.fromLTRB(0, 5, 0, 0),
+                                  child: ConstrainedBox(
+                                    constraints:BoxConstraints.tight(Size(350, 60)),
+                                    child: TextFormField(
+                                      decoration: InputDecoration(
+                                          fillColor: Colors.grey[900],
+                                          filled: true,
+                                          focusedBorder: OutlineInputBorder(
+                                            borderRadius:
+                                                BorderRadius.circular(5.0),
+                                            borderSide: BorderSide(
+                                              color: (Colors.amber[600])!,
+                                              width: 3.0,
+                                            ),
+                                          ),
+                                          enabledBorder: OutlineInputBorder(
+                                            borderRadius:
+                                                BorderRadius.circular(5.0),
+                                            borderSide: BorderSide(
+                                              color: (Colors.grey[900])!,
+                                              width: 3.0,
+                                            ),
+                                          ),
+                                          border: UnderlineInputBorder(),
+                                          hintText:
+                                              'Enter the amount of AX to stake'),
+                                    ),
+                                  ),
+                                ),
+
+                                // Approve button
+                                Padding(
+                                    padding: EdgeInsets.fromLTRB(0, 0, 0, 0),
+                                    child: SizedBox(
+                                      child: ElevatedButton(
+                                        child: const Text('APPROVE'),
+                                        onPressed: () {
+                                          showDialog(
+                                            context: context,
+                                            builder: (BuildContext context) => _buildPopupDialog(context),
+                                          );},
+                                        style: approveButton,
+                                      ),
+                                      height: 50,
+                                      width: 350,
+                                    )),
+                                
+                                Padding(
+                                    padding: EdgeInsets.fromLTRB(0, 30, 0, 0),
+                                    child: SizedBox(
+                                      child: ElevatedButton(
+                                        child: const Text('Connect to a wallet'),
+                                        onPressed: () {
+                                          showDialog(
+                                            context: context,
+                                            builder: (BuildContext context) => _buildPopupDialog(context),
+                                          );
+                                        },
+                                        style: connectButton,
+                                      ),
+                                      height: 50,
+                                      width: 350,
+                                    ))
+                              ],
+                            )),
+                          ],
+                        ))
+                      ],
+                    )),
+              ],
+            )
+          ],
+        )),
       ],
     ));
   }
